folly/MPMCQueue sizeGuess should return a signed int
Summary:
Like size(), sizeGuess() is allowed to return a negative value,
so it should be ssize_t instead of uint64_t.

Test Plan: fbmake, wait for contbuild

Reviewed By: tudorb@fb.com

FB internal diff: D893075

+++ folly/MPMCQueue.h
@@ -196,7 +196,7 @@ class MPMCQueue : boost::noncopyable {
 
   /// Returns is a guess at size() for contexts that don't need a precise
   /// value, such as stats.
-  uint64_t sizeGuess() const noexcept {
+  ssize_t sizeGuess() const noexcept {
     return writeCount() - readCount();
   }
